% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoDETesting.R
\name{DoDETesting}
\alias{DoDETesting}
\title{Perform differential expression testing.}
\usage{
DoDETesting(
  exp_object,
  condition_var,
  comparison,
  test = "wilcox",
  replicate_var = NULL,
  method = "return",
  separator = ":",
  cells = NULL,
  genes = NULL
)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{condition_var}{The string name of the metadata column that contains the
condition that is being tested.}

\item{comparison}{Either a singular value that appears in the
\code{condition_var} column or a vector of two values that appear in the
\code{condition_var} column. If a singular value is given, cells/samples
matching that value will be compared to all other cells/samples. If a vector
of two values in given, cells/samples in the two groups will be compared.}

\item{test}{The string name of the test to use. Can be \code{"wilcox"},
\code{"wilcox_presto"}, \code{"DESeq2"}, \code{"limma"}, \code{"edgeR"}, or
\code{"glmGamPoi"}. \code{"wilcox"} by default. Note: a very basic design is
used for tests other than \code{"wilcox"} or \code{"wilcox_presto"}. To use
more complex designs, revert back to the packages from where these tests
come.}

\item{replicate_var}{The string name of the metadata column that will be used
to make replicates for pseudobulking purposes. If \code{NULL}, then
pseudobulking will not be performed. \code{NULL} by default.}

\item{method}{The string name of the method that the user wants to employ.
Can be \code{"save"}, which will save the results of the testing inside of
the given \code{exp_object}, or \code{"return"}, which will return the
results of the testing as a data frame. \code{"return"} by default.}

\item{separator}{A string value to separate the condition and the replicate
labels in the column names of the resulting matrix. It is recommended to use
a value that does not appear in any of the \code{condition_var} or
\code{replicate_var} values. \code{":"} by default.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{genes}{The string name of a singular gene of interest or a vector
containing the string names of many genes of interest.}
}
\description{
This function returns the results of testing for differentially-expressed
genes across two conditions, using various methodologies/models. The function
has pseudobulking capabilities built in and supports the following tests:
Wilcoxon Rank Sum (with and without \code{presto}), DESeq2, limma, edgeR, and
glmGamPoi.
}
