% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPseudobulkAggregation.R
\name{DoPseudobulkAggregation}
\alias{DoPseudobulkAggregation}
\title{Aggregate expression data in preparation for pseudobulk DE testing.}
\usage{
DoPseudobulkAggregation(
  exp_object,
  condition_var,
  replicate_var,
  separator = ":",
  cells = NULL,
  genes = NULL
)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{condition_var}{The string name of a variable in the object's metadata
that contains the condition labels that will be used for DE testing.}

\item{replicate_var}{The string name of a variable in the object's metadata
that contains the replicate labels. These labels are usually the sample or
library that the individual cells came from.}

\item{separator}{A string value to separate the condition and the replicate
labels in the column names of the resulting matrix. It is recommended to use
a value that does not appear in any of the \code{condition_var} or
\code{replicate_var} values. \code{":"} by default.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{genes}{The string name of a singular gene of interest or a vector
containing the string names of many genes of interest. \code{NULL} by
default.}
}
\description{
This function returns a sparse matrix in dgCMatrix format that contains the
expression data from a given experiment, aggregated for replicates (usually
samples or libraries) within DE testing conditions. The user can optionally
provide specific gene IDs and/or cell IDs to make more specific queries from
the expression data. If this is done, the function has the side effect of
creating global variables called \code{Catullus_genes} that stores the vector
of genes and \code{Catullus_cells} that stores the vector of cells.
}
