% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellGroupIDs.R
\name{GetCellGroupIDs}
\alias{GetCellGroupIDs}
\title{Get the cell IDs of a group of interest from an experiment.}
\usage{
GetCellGroupIDs(exp_object, cell_query_string = NULL)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{cell_query_string}{A string representing a logical expression that can
be used to filter for cells of a specific group. \code{NULL} by default.}
}
\description{
This function returns a vector that contains the cell IDs of a group of cells
that was specified by a query on an experiment's metadata. If no query is
made, then all cell IDs will be returned.
}
