% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDEData.R
\name{GetDEData}
\alias{GetDEData}
\title{Get differential expression data from an experiment.}
\usage{
GetDEData(exp_object, de_name)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{de_name}{The string name of differential expression testing of interest
within the varm slot of the given \code{tiledbsoma} object.}
}
\description{
This function returns the data from specified differential expression testing
in a given \code{tiledbsoma} object.
}
