% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDimRedData.R
\name{GetDimRedData}
\alias{GetDimRedData}
\title{Get the dimensional reduction data from an experiment.}
\usage{
GetDimRedData(exp_object, red_name = "umap", cells = NULL)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{red_name}{The string name of the dimensional reduction of interest. Can
be \code{"umap"} or \code{"pca"}. \code{"umap"} by default.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}
}
\description{
This function returns a data frame that contains the dimensional reduction
data from a given experiment. The user can optionally provide specific
reduction names and/or cell IDs to make more specific queries from the
data.
}
