% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExpressionData.R
\name{GetExpressionData}
\alias{GetExpressionData}
\title{Get the expression data from an experiment.}
\usage{
GetExpressionData(exp_object, genes = NULL, cells = NULL, X_slot = "counts")
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{genes}{The string name of a singular gene of interest or a vector
containing the string names of many genes of interest. \code{NULL} by
default.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{X_slot}{The string name of the X layer from which the user wants to
query expression data. Can be \code{"counts"}, \code{"data"}, or
\code{"scale_data"}. \code{"counts"} by default.}
}
\description{
This function returns a sparse matrix in dgCMatrix format that contains the
expression data from a given experiment. The user can optionally provide
specific gene IDs and/or cell IDs to make more specific queries from the
expression data. If this is done, the function has the side effect of
creating global variables called \code{Catullus_genes}, which stores a
vector of genes, and \code{Catullus_cells}, which stores a vector of cells.
}
