% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetGeneData.R
\name{GetGeneData}
\alias{GetGeneData}
\title{Get the gene data from an experiment.}
\usage{
GetGeneData(exp_object, variables = NULL, genes = NULL, id_rownames = T)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{variables}{The string name of a singular variable of interest or a
vector containing the string names of many variables of interest. \code{NULL}
by default.}

\item{genes}{The string name of a singular gene of interest or a vector
containing the string names of many genes of interest. \code{NULL} by
default.}

\item{id_rownames}{A boolean indicating whether to make the row names of the
resulting data frame the gene names. \code{TRUE} by default.}
}
\description{
This function returns a data frame that contains the gene data from a given
experiment. The user can optionally provide specific variables and/or gene
names to make more specific queries from the gene data.
}
