% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetaData.R
\name{GetMetaData}
\alias{GetMetaData}
\title{Get the metadata from an experiment.}
\usage{
GetMetaData(exp_object, variables = NULL, cells = NULL, id_rownames = T)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{variables}{The string name of a singular variable of interest or a
vector containing the string names of many variables of interest. \code{NULL}
by default.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{id_rownames}{A boolean indicating whether to make the row names of the
resulting data frame the cell IDs. \code{TRUE} by default.}
}
\description{
This function returns a data frame that contains the metadata from a given
experiment. The user can optionally provide specific variables and/or cell
IDs to make more specific queries from the metadata.
}
