% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObjectSchema.R
\name{GetObjectSchema}
\alias{GetObjectSchema}
\title{Get the schema of an experiment object.}
\usage{
GetObjectSchema(exp_object, method = "print")
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{method}{A string name of the method that the user wants to employ. Can
be \code{"print"}, which will simply print the schema. Can also be
\code{"return"}, which will return a list object that the user can then parse
through. \code{"print"} by default.}
}
\description{
This function will print or return the basic schema of a given experiment
object. Its main use is to allow users to examine the contents of an object
and to see the various names of measurements throughout it.
}
