% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewDETesting.R
\name{ViewDETesting}
\alias{ViewDETesting}
\title{View differential expression testing results.}
\usage{
ViewDETesting(
  input_object,
  de_name = NULL,
  title = NULL,
  repel_overlaps = 50,
  lfc_cutoff = 1,
  p_adj_cutoff = 0.05,
  cols = c("blue", "grey", "red")
)
}
\arguments{
\item{input_object}{A \code{tiledbsoma} experiment object or a data frame
returned by \code{DoDETesting}.}

\item{de_name}{The string name of differential expression testing of interest
within the varm slot of the given \code{tiledbsoma} object. \code{NULL} by
default.}

\item{title}{A title for the plot. \code{NULL} by default.}

\item{repel_overlaps}{An integer specifying how many genes can be labeled.
\code{50} by default.}

\item{lfc_cutoff}{The log2FC cutoff for a gene to be labeled on the plot.
\code{1} by default.}

\item{p_adj_cutoff}{The p_adj cutoff for a gene to be labeled on the plot.
\code{0.05} by default.}

\item{cols}{A vector of three colors corresponding to the how the colors will
appear on the plot, from left to right. \code{c("blue", "grey", "red")} by
default.}
}
\description{
This function returns a volcano plot, displaying the results of specified
differential expression testing within a \code{tiledbsoma} object.
}
