% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewExpression.R
\name{ViewExpression}
\alias{ViewExpression}
\title{View the expression of a gene or a gene list.}
\usage{
ViewExpression(
  exp_object,
  genes,
  cells = NULL,
  X_slot = "data",
  split_var = NULL,
  label_var = NULL,
  order_expression = T,
  colors = c("snow2", "blue"),
  pt_size = 0.2,
  label_size = 3,
  plot_style = "bare",
  method = "print"
)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{genes}{The string name of a singular gene of interest or a vector
containing the string names of many genes of interest.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{X_slot}{The string name of the X layer for which the user wants to see
expression. Can be \code{"counts"}, \code{"data"}, or \code{"scale_data"}.
\code{"data"} by default.}

\item{split_var}{The string name of the metadata column that the user wants
to split the plot by. \code{NULL} by default.}

\item{label_var}{The string name of the metadata column that the user wants
to use to label the data. \code{NULL} by default.}

\item{order_expression}{A boolean indicating whether to order the cells by
decreasing expression. \code{TRUE} by default.}

\item{colors}{A vector of a low color and a high color, in that order,
to match the expression values to. \code{"snow2"} and \code{"blue"} by
default.}

\item{pt_size}{A float indicating the size of each point. 0.2 by default.}

\item{label_size}{A float indicating the size the labels. \code{3} by
default.}

\item{plot_style}{A string name of the plot style that the user wants. Can be
\code{"bare"}, \code{"standard"}, or \code{"detailed"}. \code{"bare"} by
default.}

\item{method}{A string name of the method that the user wants to employ. Can
be \code{"print"}, which will simply print the plot or plots. Can also be
\code{"return"}, which will return a list of plots that the user can then
interact with later. \code{"print"} by default.}
}
\description{
This function returns a plot in which a gene's expression values appear for
every cell on an experiment's UMAP. The function can handle singular genes or
a vector of genes. the function has the side effect of creating global
variables called \code{Catullus_genes}, which stores a vector of genes, and
\code{Catullus_cells}, which stores a vector of cells.
}
