% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewMeta.R
\name{ViewMeta}
\alias{ViewMeta}
\title{View one metadata variable or many.}
\usage{
ViewMeta(
  exp_object,
  color_var,
  cells = NULL,
  labeling = F,
  legend = T,
  split_var = NULL,
  shape_var = NULL,
  pt_size = 0.2,
  label_size = 3,
  plot_style = "bare"
)
}
\arguments{
\item{exp_object}{A \code{tiledbsoma} experiment object.}

\item{color_var}{The string name of the metadata column that the user wants
to color the plot by.}

\item{cells}{The string name of a singular cell of interest or a vector
containing the string names of many cells of interest. \code{NULL} by
default.}

\item{labeling}{A boolean indicating whether to label the cells by the levels
in the metadata column that the plot is colored by. \code{FALSE} by default.}

\item{legend}{A boolean indicating whether to include a legend for the levels
in the metadata column that the plot is colored by. \code{TRUE} by default.}

\item{split_var}{The string name of the metadata column that the user wants
to split the plot by. \code{NULL} by default.}

\item{shape_var}{The string name of the metadata column that the user wants
to shape the plot by. \code{NULL} by default.}

\item{pt_size}{A float indicating the size of each point. \code{0.2} by
default.}

\item{label_size}{A float indicating the size the labels. \code{3} by
default.}

\item{plot_style}{A string name of the plot style that the user wants. Can be
\code{"bare"}, \code{"standard"}, or \code{"detailed"}. \code{"bare"} by
default.}
}
\description{
This function returns a plot in which each cell's status across one or more
metadata variables is shown on an experiment's UMAP. The function can color,
shape, and split cells based on metadata variables from the experiment.
}
